# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AssistantProfile {
  id: string;
  email: string;
  displayName: string;
  preferredLanguage: string;
  profession: string;
  project: string;
  plan: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface AssistantProfileUpdate {
  displayName?: string;
  preferredLanguage?: string;
  profession?: string;
  project?: string;
}

export interface AssistantChat {
  id: string;
  title: string;
  updatedAt: string;
  messageCount: number;
}

export interface AssistantChatInput {
  title?: string;
}

export type AssistantMessageRole = typeof AssistantMessageRole[keyof typeof AssistantMessageRole];


export const AssistantMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AssistantMessage {
  id: string;
  chatId: string;
  role: AssistantMessageRole;
  content: string;
  createdAt: string;
}

export interface AssistantMessageInput {
  /** @minLength 1 */
  content: string;
}

export type AssistantMemoryItemCategory = typeof AssistantMemoryItemCategory[keyof typeof AssistantMemoryItemCategory];


export const AssistantMemoryItemCategory = {
  meeting: 'meeting',
  code: 'code',
  note: 'note',
} as const;

export interface AssistantMemoryItem {
  id: string;
  category: AssistantMemoryItemCategory;
  title: string;
  content: string;
  updatedAt: string;
}

export type AssistantMemoryInputCategory = typeof AssistantMemoryInputCategory[keyof typeof AssistantMemoryInputCategory];


export const AssistantMemoryInputCategory = {
  meeting: 'meeting',
  code: 'code',
  note: 'note',
} as const;

export interface AssistantMemoryInput {
  category: AssistantMemoryInputCategory;
  title: string;
  content: string;
}