# lib/api-zod/src/generated/types/assistantChat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantChat {
  id: string;
  title: string;
  updatedAt: string;
  messageCount: number;
}