# lib/api-zod/src/generated/types/assistantMemoryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantMemoryInputCategory } from './assistantMemoryInputCategory';

export interface AssistantMemoryInput {
  category: AssistantMemoryInputCategory;
  title: string;
  content: string;
}