# lib/api-zod/src/generated/types/assistantMemoryItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantMemoryItemCategory } from './assistantMemoryItemCategory';

export interface AssistantMemoryItem {
  id: string;
  category: AssistantMemoryItemCategory;
  title: string;
  content: string;
  updatedAt: string;
}