# lib/api-zod/src/generated/types/assistantProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantProfile {
  id: string;
  email: string;
  displayName: string;
  preferredLanguage: string;
  profession: string;
  project: string;
  plan: string;
  /** @nullable */
  avatarUrl?: string | null;
}